:sparkles: Add interractive button for Wishlist
As you can see, this has nothing to do with the Dropdown, in the  future I will choose or make a new branch accordingly and code more carefully. Now i just want to finish with the main functionality as soon as possible, due the deadline I have. Afterwards I will refactor and do it more properly. .